optimize(core): move some tool functions into tokenizer
